Add files via upload
i addid api useing fastapi and I made files for students so you can add delet or view students in the api also did some changing in the code so it can be easy to read and fix small issues that occurred to me when I was trying, like I needed to change the dict was the problem since it couldn't understand what type of file in the dictionary. I needed to identify each type i also I also explored Swagger/OpenAPI documentation, request/response models, and error handling for clean, production-ready APIs.